backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "attrs==25.1.0",
    "bcrypt==4.2.1",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "coverage==7.6.12",
    "cryptography==44.0.0",
    "dnspython==2.7.0",
    "docx2txt==0.9",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi-cli==0.0.5",
    "fastapi[standard]==0.115.4",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "langchain>=0.3.24",
    "langdetect==1.0.9",
    "llama-index>=0.12.33",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "neo4j>=5.28.1",
    "nltk==3.9.1",
    "numpy==2.1.3",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pandas==2.2.3",
    "passlib==1.7.4",
    "pillow==11.0.0",
    "pluggy==1.5.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-settings==2.8.1",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pymupdf==1.24.13",
    "pytesseract==0.3.13",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.17",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "reportlab==4.3.1",
    "requests==2.32.3",
    "rich==13.9.4",
    "rsa==4.9",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.36",
    "sqlmodel==0.0.22",
    "starlette==0.41.2",
    "threadpoolctl==3.5.0",
    "tqdm==4.67.0",
    "trio==0.29.0",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "unidecode==1.3.8",
    "uuid>=1.30",
    "uvicorn==0.32.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
]
